feat: ✨ add Zod to JSON schema conversion helpers
